Time: 0 ms (100.00%), Space: 5.9 MB (88.86%) - LeetHub

+++ 0062-unique-paths/0062-unique-paths.cpp
@@ -5,10 +5,8 @@ public:
         int r = n-1;
         double ans = 1;
         for(int i = 1;i<=r;i++){
-            ans = ans * (N - r + i) / i;
-            // ans = ans/i+1;
+            ans = ans * (N - r + i)/ i;
         }
-        return (int)ans;
-        // return 0;
+        return ans;
     }
 };
